feat: implementation of error handling and exceptions

+++ src/modules/event/event.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common'
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common'
 import { CreateEventDto } from './dto/create-event.dto'
 import { UpdateEventDto } from './dto/update-event.dto'
 import { EventRepository } from './repository/event.repository'
@@ -11,23 +11,51 @@ export class EventService {
       @InjectRepository(Event)
       private repository: EventRepository,
    ) {}
-   create(createEventDto: CreateEventDto) {
-      return this.repository.create(createEventDto)
-   }
 
-   findAll() {
+   findAll(): Promise<Event[]> {
       return this.repository.find()
    }
 
-   findOne(id: number) {
+   async findOne(id: number): Promise<Event> {
+      try {
+         const event = await this.repository.findOneById(id)
+
+         if (event === null) {
+            throw new HttpException('Evento não encontrado.', HttpStatus.NOT_FOUND)
+         }
+
+         return event
+      } catch (error) {
+         throw error
+      }
+   }
+
+   async create(createEventDto: CreateEventDto): Promise<Event> {
+      const newEvent = await this.repository.create(createEventDto)
+
+      await this.repository.save(newEvent)
+      return newEvent
+   }
+
+   async update(id: number, updateEventDto: UpdateEventDto): Promise<UpdateEventDto> {
+      await this.repository.update(id, updateEventDto)
       return this.repository.findOneById(id)
    }
 
-   update(id: number, updateEventDto: UpdateEventDto) {
-      return this.repository.update(id, updateEventDto)
-   }
+   async remove(id: number): Promise<void> {
+      try {
+         const event = await this.repository.findOneById(id)
 
-   remove(id: number) {
-      return this.repository.delete(id)
+         if (event === null) {
+            throw new HttpException(
+               'Não foi possível deletar evento.',
+               HttpStatus.BAD_REQUEST,
+            )
+         }
+
+         await this.repository.delete(id)
+      } catch (error) {
+         throw error
+      }
    }
 }
